Extracted file_index_type into its own source files
