+++ vme/src/dbfind.cpp
@@ -8,6 +8,7 @@
 
 #include "db.h"
 #include "dil.h"
+#include "file_index_type.h"
 #include "handler.h"
 #include "main_functions.h"
 #include "structs.h"
